fitting for physical, with fake error bars
