+++ setup.py
@@ -58,8 +58,12 @@ def find_scripts():
 
     sdir = 'scripts'
 
-    return [os.path.join(sdir, val) for val in os.listdir(sdir) if
-            val.endswith('.py') and '__init__' not in val]
+    res = [
+        os.path.join(sdir, val) for val in os.listdir(sdir) if
+            val.endswith('.py') and '__init__' not in val
+    ]
+
+    return res
 
 
 setup(
